Move SDRConfig out of SDRDevice
