+++ src/boards/LMS7002M_SDRDevice.h
@@ -130,10 +130,10 @@ class LIME_API LMS7002M_SDRDevice : public SDRDevice
     static OpStatus UpdateFPGAInterfaceFrequency(LMS7002M& soc, FPGA& fpga, uint8_t chipIndex);
     void SetGainInformationInDescriptor(RFSOCDescriptor& descriptor);
 
-    OpStatus LMS7002LOConfigure(LMS7002M* chip, const SDRDevice::SDRConfig& config);
-    OpStatus LMS7002ChannelConfigure(LMS7002M* chip, const SDRDevice::ChannelConfig& config, uint8_t channelIndex);
-    OpStatus LMS7002ChannelCalibration(LMS7002M* chip, const SDRDevice::ChannelConfig& config, uint8_t channelIndex);
-    OpStatus LMS7002TestSignalConfigure(LMS7002M* chip, const SDRDevice::ChannelConfig& config, uint8_t channelIndex);
+    OpStatus LMS7002LOConfigure(LMS7002M* chip, const SDRConfig& config);
+    OpStatus LMS7002ChannelConfigure(LMS7002M* chip, const ChannelConfig& config, uint8_t channelIndex);
+    OpStatus LMS7002ChannelCalibration(LMS7002M* chip, const ChannelConfig& config, uint8_t channelIndex);
+    OpStatus LMS7002TestSignalConfigure(LMS7002M* chip, const ChannelConfig& config, uint8_t channelIndex);
 
     static constexpr uint8_t NCOValueCount = 16;
