+++ src/boards/LimeSDR_X3/LimeSDR_X3.cpp
@@ -648,7 +648,7 @@ OpStatus LimeSDR_X3::Configure(const SDRConfig& cfg, uint8_t socIndex)
 void LimeSDR_X3::ConfigureDirection(TRXDir dir, LMS7002M* chip, const SDRConfig& cfg, int ch, uint8_t socIndex)
 {
     const char* dirName = dir == TRXDir::Rx ? "Rx" : "Tx";
-    SDRDevice::ChannelConfig::Direction trx = cfg.channel[ch].GetDirection(dir);
+    ChannelConfig::Direction trx = cfg.channel[ch].GetDirection(dir);
 
     if (socIndex == 1) // LMS2 uses external ADC/DAC
     {
@@ -715,7 +715,7 @@ void LimeSDR_X3::ConfigureDirection(TRXDir dir, LMS7002M* chip, const SDRConfig&
     }
 }
 
-void LimeSDR_X3::SetLMSPath(const TRXDir dir, const SDRDevice::ChannelConfig::Direction& trx, const int ch, const uint8_t socIndex)
+void LimeSDR_X3::SetLMSPath(const TRXDir dir, const ChannelConfig::Direction& trx, const int ch, const uint8_t socIndex)
 {
     switch (socIndex)
     {
