+++ src/boards/LimeSDR_X3/LimeSDR_X3.h
@@ -78,7 +78,7 @@ class LimeSDR_X3 : public LMS7002M_SDRDevice
 
   private:
     void ConfigureDirection(TRXDir dir, LMS7002M* chip, const SDRConfig& cfg, int ch, uint8_t socIndex);
-    void SetLMSPath(const TRXDir dir, const SDRDevice::ChannelConfig::Direction& trx, const int ch, const uint8_t socIndex);
+    void SetLMSPath(const TRXDir dir, const ChannelConfig::Direction& trx, const int ch, const uint8_t socIndex);
 
     CDCM_Dev* mClockGeneratorCDCM;
     Equalizer* mEqualizer;
