+++ src/examples/basicRX.cpp
@@ -62,7 +62,7 @@ int main(int argc, char** argv)
     device->Init();
 
     // RF parameters
-    SDRDevice::SDRConfig config;
+    SDRConfig config;
     config.channel[0].rx.enabled = true;
     config.channel[0].rx.centerFrequency = frequencyLO;
     config.channel[0].rx.sampleRate = sampleRate;
