+++ src/lms7002m/LMS7002M_validation.cpp
@@ -15,12 +15,12 @@ static inline bool InRange(double val, double min, double max)
     return val >= min ? val <= max : false;
 }
 
-bool LMS7002M_Validate(const SDRDevice::SDRConfig& cfg, std::vector<std::string>& errors, const int lms7002_chCount)
+bool LMS7002M_Validate(const SDRConfig& cfg, std::vector<std::string>& errors, const int lms7002_chCount)
 {
     errors.clear();
 
     // channel count check
-    for (int i = lms7002_chCount; i < SDRDevice::MAX_CHANNEL_COUNT; ++i)
+    for (int i = lms7002_chCount; i < SDRConfig::MAX_CHANNEL_COUNT; ++i)
     {
         if (cfg.channel[i].rx.enabled)
             errors.push_back(strFormat("Rx channel %i enabled, LMS7002M only has %i channels", i, lms7002_chCount));
@@ -29,8 +29,8 @@ bool LMS7002M_Validate(const SDRDevice::SDRConfig& cfg, std::vector<std::string>
     }
 
     // MIMO necessary checks
-    const SDRDevice::ChannelConfig& chA = cfg.channel[0];
-    const SDRDevice::ChannelConfig& chB = cfg.channel[1];
+    const ChannelConfig& chA = cfg.channel[0];
+    const ChannelConfig& chB = cfg.channel[1];
     const bool rxMIMO = chA.rx.enabled && chB.rx.enabled;
     const bool txMIMO = chA.tx.enabled && chB.tx.enabled;
     if (rxMIMO || txMIMO)
@@ -60,7 +60,7 @@ bool LMS7002M_Validate(const SDRDevice::SDRConfig& cfg, std::vector<std::string>
     const double maxLO = 3.8e9;
     for (int i = 0; i < 2; ++i)
     {
-        const SDRDevice::ChannelConfig& ch = cfg.channel[i];
+        const ChannelConfig& ch = cfg.channel[i];
         if (ch.rx.enabled && !InRange(ch.rx.centerFrequency, minLO, maxLO))
             errors.push_back(strFormat("Rx ch%i LO (%g) out of range [%g:%g]", i, ch.rx.centerFrequency, minLO, maxLO));
         if (ch.tx.enabled && !InRange(ch.tx.centerFrequency, minLO, maxLO))
