+++ src/lms7002m/LMS7002M_validation.h
@@ -4,7 +4,7 @@
 namespace lime {
 
 // Checks if config parameters are within expected values and can produce working state
-bool LMS7002M_Validate(const SDRDevice::SDRConfig& config, std::vector<std::string>& out_errors, const int lms7002_chCount = 2);
+bool LMS7002M_Validate(const SDRConfig& config, std::vector<std::string>& out_errors, const int lms7002_chCount = 2);
 
 } // namespace lime
